backend/tests/test_images.py
"""
图像模块API接口测试
使用api_client测试运行中的后端服务
"""
import pytest
import base64
import io
from PIL import Image


BASE_URL = "http://localhost:8000"


def create_test_image():
    """创建测试图像"""
    img = Image.new("RGB", (100, 100), color="red")
    buffer = io.BytesIO()
    img.save(buffer, format="JPEG")
    buffer.seek(0)
    return buffer


class TestImageUpload:
    """测试图像上传接口"""

    def test_upload_single_image_success(self, api_client, auth_token):
        """测试成功上传单张图像"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        img_data = create_test_image()
        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/upload",
            files={"file": ("test.jpg", img_data, "image/jpeg")},
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[POST] /api/v1/images/images/upload - Status: {response.status_code}")
        print(f"Response: {response.json()}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") == 0

    def test_upload_image_no_file(self, api_client, auth_token):
        """测试上传未选择文件"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/upload",
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[POST] /api/v1/images/images/upload (no file) - Status: {response.status_code}")
        # 参数验证错误可能返回422或200带errcode
        assert response.status_code in [200, 422]
        if response.status_code == 200:
            data = response.json()
            assert data.get("errcode") != 0

    def test_upload_image_invalid_format(self, api_client, auth_token):
        """测试上传无效格式"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        text_data = io.BytesIO(b"This is not an image")
        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/upload",
            files={"file": ("test.txt", text_data, "text/plain")},
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[POST] /api/v1/images/images/upload (invalid format) - Status: {response.status_code}")
        # 返回200但errcode不为0表示业务错误
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") != 0

    def test_upload_batch_success(self, api_client, auth_token):
        """测试成功批量上传图像"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        files = []
        for i in range(3):
            img_data = create_test_image()
            files.append(("files", (f"test{i}.jpg", img_data, "image/jpeg")))

        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/upload/batch",
            files=files,
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[POST] /api/v1/images/images/upload/batch - Status: {response.status_code}")
        print(f"Response: {response.json()}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") == 0

    def test_upload_batch_too_many_images(self, api_client, auth_token):
        """测试批量上传超过限制"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        files = []
        for i in range(10):  # 超过9张限制
            img_data = create_test_image()
            files.append(("files", (f"test{i}.jpg", img_data, "image/jpeg")))

        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/upload/batch",
            files=files,
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[POST] /api/v1/images/images/upload/batch (too many) - Status: {response.status_code}")
        # 返回200但errcode不为0表示业务错误
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") != 0

    def test_upload_without_auth(self, api_client):
        """测试未认证上传图像"""
        img_data = create_test_image()
        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/upload",
            files={"file": ("test.jpg", img_data, "image/jpeg")}
        )
        print(f"\n[POST] /api/v1/images/images/upload (no auth) - Status: {response.status_code}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") == 2001  # 未认证


class TestGetImages:
    """测试获取图像列表接口"""

    def test_get_images_with_auth(self, api_client, auth_token):
        """测试带认证获取图像列表"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.get(
            f"{BASE_URL}/api/v1/images/images",
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[GET] /api/v1/images/images - Status: {response.status_code}")
        print(f"Response: {response.json()}")
        assert response.status_code == 200
        data = response.json()
        # API可能返回内部错误(4001)如果数据库连接问题
        assert data.get("errcode") in [0, 4001]
        if data.get("errcode") == 0:
            assert data.get("data") is not None
            assert "images" in data.get("data", {})
            assert "total" in data.get("data", {})

    def test_get_images_pagination(self, api_client, auth_token):
        """测试图像列表分页"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.get(
            f"{BASE_URL}/api/v1/images/images?page=1&per_page=10",
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[GET] /api/v1/images/images?page=1&per_page=10 - Status: {response.status_code}")
        assert response.status_code == 200
        data = response.json()
        # API可能返回内部错误(4001)如果数据库连接问题
        assert data.get("errcode") in [0, 4001]

    def test_get_images_without_auth(self, api_client):
        """测试未认证获取图像列表"""
        response = api_client.get(f"{BASE_URL}/api/v1/images/images")
        print(f"\n[GET] /api/v1/images/images (no auth) - Status: {response.status_code}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") == 2001  # 未认证


class TestGetImageDetail:
    """测试获取图像详情接口"""

    def test_get_image_detail_success(self, api_client, auth_token):
        """测试成功获取图像详情"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.get(
            f"{BASE_URL}/api/v1/images/images/detail?image_id=1",
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[GET] /api/v1/images/images/detail?image_id=1 - Status: {response.status_code}")
        print(f"Response: {response.json()}")
        assert response.status_code == 200
        data = response.json()
        # 图像可能不存在，但响应格式应该正确

    def test_get_image_detail_not_found(self, api_client, auth_token):
        """测试获取不存在的图像详情"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.get(
            f"{BASE_URL}/api/v1/images/images/detail?image_id=99999",
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[GET] /api/v1/images/images/detail?image_id=99999 - Status: {response.status_code}")
        print(f"Response: {response.json()}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") != 0

    def test_get_image_detail_missing_id(self, api_client, auth_token):
        """测试获取图像详情缺少ID"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.get(
            f"{BASE_URL}/api/v1/images/images/detail",
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[GET] /api/v1/images/images/detail (no id) - Status: {response.status_code}")
        # API可能返回200带错误码或422
        assert response.status_code in [200, 422]
        if response.status_code == 200:
            data = response.json()
            assert data.get("errcode") != 0

    def test_get_image_detail_without_auth(self, api_client):
        """测试未认证获取图像详情"""
        response = api_client.get(f"{BASE_URL}/api/v1/images/images/detail?image_id=1")
        print(f"\n[GET] /api/v1/images/images/detail (no auth) - Status: {response.status_code}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") == 2001  # 未认证


class TestDeleteImage:
    """测试删除图像接口"""

    def test_delete_image_success(self, api_client, auth_token):
        """测试成功删除图像"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/delete",
            json={"image_id": 99999},  # 使用不存在的ID避免误删
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[POST] /api/v1/images/images/delete - Status: {response.status_code}")
        print(f"Response: {response.json()}")
        assert response.status_code == 200
        data = response.json()
        # 图像不存在，errcode不为0

    def test_delete_image_not_found(self, api_client, auth_token):
        """测试删除不存在的图像"""
        if auth_token is None:
            pytest.skip("Login failed, cannot test authenticated endpoint")

        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/delete",
            json={"image_id": 99999},
            headers={"Authorization": f"Bearer {auth_token}"}
        )
        print(f"\n[POST] /api/v1/images/images/delete (not found) - Status: {response.status_code}")
        print(f"Response: {response.json()}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") != 0

    def test_delete_image_without_auth(self, api_client):
        """测试未认证删除图像"""
        response = api_client.post(
            f"{BASE_URL}/api/v1/images/images/delete",
            json={"image_id": 1}
        )
        print(f"\n[POST] /api/v1/images/images/delete (no auth) - Status: {response.status_code}")
        assert response.status_code == 200
        data = response.json()
        assert data.get("errcode") == 2001  # 未认证


if __name__ == "__main__":
    pytest.main([__file__, "-v", "-s"])
