backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1

# 异步数据库驱动
asyncpg==0.29.0  # PostgreSQL异步驱动
aiomysql==0.2.0  # MySQL异步驱动
aiosqlite==0.19.0  # SQLite异步驱动

psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI模型依赖 - 本地模型已移除，使用 ModelScope API
# 如需本地部署，取消以下注释:
# torch>=2.0.0
# torchvision>=0.15.0
# transformers==4.36.0
# diffusers==0.24.0
# accelerate==0.25.0
# xformers==0.0.22
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4

# 图像处理 - 精简版 (本地模型移除后)
imageio==2.31.1
scikit-image==0.22.0
# rembg==2.0.52  # 可选，需要本地模型时取消注释

# 文字检测和处理 - 精简版 (本地模型移除后)
# easyocr==1.7.0  # 可选，需要本地 OCR 时取消注释
pytesseract==0.3.10
pdf2image==1.17.0

# 数据验证和处理
marshmallow==3.20.2
pydantic[email]==2.5.0
pandas==2.1.4

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# 文件处理
aiofiles==23.2.1
python-magic==0.4.27

# 监控和日志
sentry-sdk[fastapi]==1.38.0
loguru==0.7.2

# 邮件
fastapi-mail==1.4.1

# 支付
stripe==7.8.0

# 云存储
boto3==1.34.0
minio==7.2.0

# API文档
fastapi-limiter==0.1.6

# 工具
click==8.1.7
rich==13.7.0
typer==0.9.0

# 安全
cryptography==41.0.8
validators==0.22.0

# 缓存
cachetools==5.3.2

# 定时任务
apscheduler==3.10.4

# 测试覆盖率
pytest-cov==4.1.0